requirements.txt
# ============================================================================
# Granite-8B Fine-Tuning Dependencies
# ============================================================================
# For embedded automotive code generation (AVB/TSN)
# ============================================================================

# Core ML libraries - UPDATED FOR GRANITE-8B
torch>=2.1.0
transformers>=4.45.2
datasets==2.14.5
accelerate>=0.31.0
peft>=0.12.0
trl>=0.9.6
bitsandbytes>=0.43.0

# SentencePiece for tokenization
sentencepiece>=0.1.99
protobuf>=3.20.0

# AWS SDK
boto3>=1.34.0
sagemaker>=2.199.0

# Data processing
pandas>=2.1.0
numpy>=1.24.3
jsonlines>=4.0.0
pyarrow==14.0.1  # Compatibility with datasets 2.14.5

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Progress and formatting
tqdm>=4.66.1
colorama>=0.4.6
rich>=13.6.0

# Evaluation metrics
rouge-score>=0.1.2
sacrebleu>=2.3.1
evaluate>=0.4.0

# Scipy for numerical operations
scipy>=1.11.0

# Development tools
pytest>=7.4.2
black>=23.9.1
flake8>=6.1.0

# Optional: Weights & Biases logging
# wandb>=0.15.12

# Optional: Jupyter for notebooks
# jupyter>=1.0.0
# ipykernel>=6.25.0

# ============================================================================
# Context Database (Vectorized Knowledge Graph)
# ============================================================================
neo4j>=5.14.0
sentence-transformers>=2.2.2
keybert>=0.8.3
mcp>=1.0.0
